Solution

+++ app/main.py
@@ -5,12 +5,9 @@ class CleanUpFile:
     def __init__(self, filename: str) -> None:
         self.filename = filename
 
-    def __enter__(self) -> None:
+    def __enter__(self) -> "CleanUpFile":
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb) -> None:
-        try:
-            with open(self.filename, "r"):
-                os.remove(self.filename)
-        except FileNotFoundError:
-            pass
+    def __exit__(self, exc_type: None, exc_val: None, traceback: None) -> None:
+        if os.path.exists(self.filename):
+            os.remove(self.filename)
